Chairman
Today, we are here to say thank you to those who have sacrificed to being part of this great movement. We are currently fighting the dynasties who have refused to let go of the ruler ship of this Great Nation.
These are the forces that campaigned for the constitution, and are now seeking to change the same constitution to benefit their selfish interests whose names are well known.
The same forces who have benefited from land injustices and all manner of depressing the hustler nation are trying to hood ink us that they mean well for this country Kenya.
So as we focus on the fight ahead, the central question that must be in our minds is, “What next must FORM do for Kenya?” We must therefore issue a new call to action for our beloved nation. We must not live in the memory of our past victories but join hands and roll up our sleeves again and begin a new push for the final liberation of Kenya.
In FORM, we believe in the people and the power of the people.We continue to cherish all the Friends of Ruto Movement members and we will strengthen it. With your mandate, we will invite more members to join us. As we embark on the journey for the next general elections, we must mobilize and push new frontiers of good governance.
We have a landmark election coming next. We must mobilise to win it.
Then we will deal a death blow to the DYNASTIES that is regrouping and end the culture of impunity
Let then each put in more efforts to bring new converts ready to see that His Excellency Dr. Ruto wins the presidential race.

Let’s tell our brothers and sisters, fathers and mothers out there that if they believe in defending country, they belong in FORM.
If they believe in equity, they belong in FORM.
If they believe in inclusivity, they belong in FORM.
If they believe in justice, they belong in FORM.
If they believe in a nation in which no one goes to bed in an empty stomach, they belong in FORM.
If they believe that devolution is the way to go, that counties must be protected and not undermined, FORM is the place for them.
Let us tell our people to get out of Pharaoh’s palace where they are currently feasting and dash for freedom. And let us summon the courage and tell the Pharaoh, “let our people go.”
Thank you. God Bless FORM. God bless Kenya.